feat: Redefined factory
